Revert "Serialized the AP's jump into stage 2"
This reverts commit be254a5483a1bb79b287eaf3d6b21a227c445989.

+++ crates/stage_two_abi/src/lib.rs
@@ -57,7 +57,7 @@ impl Manifest {
         F: Fn(&str) -> Option<usize>,
     {
         // Find manifest
-        let manifest = find_symbol("__manifest")?;
+        let manifest = find_symbol("__manifest")? as usize;
         let manifest = &mut *(manifest as *mut Manifest);
 
         Some(manifest)
